Refactor coredata into CQRS
This simplifies our code a little and removes an empty package.
