+++ pkg/cqrs/ddb/execution.go
@@ -4,22 +4,12 @@ import (
 	"context"
 	"encoding/json"
 
-	"github.com/inngest/inngest/pkg/coredata"
-	"github.com/inngest/inngest/pkg/cqrs"
 	"github.com/inngest/inngest/pkg/inngest"
 )
 
-func NewExecutionLoader(w cqrs.Manager) coredata.ExecutionLoader {
-	return executionLoader{w}
-}
-
-type executionLoader struct {
-	cqrs.Manager
-}
-
 // Functions returns all functions as inngest functions.
-func (el executionLoader) Functions(ctx context.Context) ([]inngest.Function, error) {
-	all, _ := el.GetFunctions(ctx)
+func (w wrapper) Functions(ctx context.Context) ([]inngest.Function, error) {
+	all, _ := w.GetFunctions(ctx)
 	funcs := make([]inngest.Function, len(all))
 	for n, i := range all {
 		f := inngest.Function{}
@@ -30,9 +20,9 @@ func (el executionLoader) Functions(ctx context.Context) ([]inngest.Function, er
 }
 
 // FunctionsScheduled returns all scheduled functions available.
-func (el executionLoader) FunctionsScheduled(ctx context.Context) ([]inngest.Function, error) {
+func (w wrapper) FunctionsScheduled(ctx context.Context) ([]inngest.Function, error) {
 	// TODO: Make less naive by storing triggers and caching.
-	fns, err := el.Functions(ctx)
+	fns, err := w.Functions(ctx)
 	if err != nil {
 		return nil, err
 	}
@@ -49,9 +39,9 @@ func (el executionLoader) FunctionsScheduled(ctx context.Context) ([]inngest.Fun
 }
 
 // FunctionsByTrigger returns functions for the given trigger by event name.
-func (el executionLoader) FunctionsByTrigger(ctx context.Context, eventName string) ([]inngest.Function, error) {
+func (w wrapper) FunctionsByTrigger(ctx context.Context, eventName string) ([]inngest.Function, error) {
 	// TODO: Make less naive by storing triggers and caching.
-	fns, err := el.Functions(ctx)
+	fns, err := w.Functions(ctx)
 	if err != nil {
 		return nil, err
 	}
