+++ pkg/cqrs/execution.go
@@ -1,4 +1,4 @@
-package coredata
+package cqrs
 
 import (
 	"context"
@@ -6,20 +6,9 @@ import (
 	"github.com/inngest/inngest/pkg/inngest"
 )
 
-/*
-type ReadWriter interface {
-	APIReadWriter
-	ExecutionLoader
-}
-*/
-
 // ExecutionLoader is an interface which specifies all functions required to run
 // workers and executors.
 type ExecutionLoader interface {
-	ExecutionFunctionLoader
-}
-
-type ExecutionFunctionLoader interface {
 	// Functions returns all functions.
 	Functions(ctx context.Context) ([]inngest.Function, error)
 	// FunctionsScheduled returns all scheduled functions available.
